Rename folder "LocalSQL" to "MySQLDriver"
