+++ TiFBackendUtils/TiFUserUtils/TiFUserResponse.ts
@@ -1,4 +1,4 @@
-import { MySQLExecutableDriver } from "../LocalSQL/MySQLDriver.js"
+import { MySQLExecutableDriver } from "../MySQLDriver/MySQLDriver.js"
 import { DBuser } from "../entities.js"
 import { ExtractSuccess } from "../result.js"
 import { UserRelationship } from "./UserRelationships.js"
